backend/pyproject.toml
[project]
name = "dog-sports-sso"
version = "0.1.0"
description = ""
authors = [
    {name = "smd67",email = "dimigstephenm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright (>=1.56.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "pymysql[rsa] (>=1.1.2,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
